Implements set and card editing
Adds functionality to edit existing sets and cards.

This commit introduces a new menu option that allows users
to either rename a set or modify the content of a specific card
within a set. The user is prompted to select the set or card
they wish to edit, and then can input the new name or content.